[Behat] Extract ResettingPasswordContext out of LoginContext
