+++ src/Sylius/Behat/Resources/config/services/contexts/ui.xml
@@ -28,8 +28,6 @@
         <service id="sylius.behat.context.ui.admin.login" class="Sylius\Behat\Context\Ui\Admin\LoginContext">
             <argument type="service" id="sylius.behat.page.admin.dashboard" />
             <argument type="service" id="sylius.behat.page.admin.login" />
-            <argument type="service" id="sylius.behat.page.admin.request_password_reset" />
-            <argument type="service" id="sylius.behat.notification_checker"/>
         </service>
 
         <service id="sylius.behat.context.ui.admin.managing_administrators" class="Sylius\Behat\Context\Ui\Admin\ManagingAdministratorsContext">
@@ -317,6 +315,11 @@
             <argument type="service" id="sylius.behat.page.admin.taxon.create" />
             <argument type="service" id="sylius.behat.notification_checker" />
             <tag name="fob.context_service" />
+        </service>
+
+        <service id="sylius.behat.context.ui.admin.resetting_password" class="Sylius\Behat\Context\Ui\Admin\ResettingPasswordContext">
+            <argument type="service" id="sylius.behat.page.admin.request_password_reset" />
+            <argument type="service" id="sylius.behat.notification_checker" />
         </service>
 
         <service id="sylius.behat.context.ui.browser" class="Sylius\Behat\Context\Ui\BrowserContext">
